Feat: move functions about cpu kv_store in worker.py to worker_base
