+++ vllm/worker/worker_base.py
@@ -14,7 +14,7 @@ from vllm.model_executor.layers.sampler import SamplerOutput
 from vllm.platforms import current_platform
 from vllm.sequence import (ExecuteModelRequest, IntermediateTensors,
                            SequenceGroupMetadata)
-from vllm.store.kv_store import KVStoreMeta
+from vllm.store.kv_store import BlockMappingFromCPU, KVBlockStore, KVStoreMeta
 from vllm.utils import (enable_trace_function_call_for_thread,
                         resolve_obj_by_qualname, update_environment_variables)
 from vllm.worker.model_runner_base import (BroadcastableModelInput,
@@ -33,6 +33,7 @@ class WorkerBase(ABC):
     def __init__(
         self,
         vllm_config: VllmConfig,
+        local_rank: int = 0,
     ) -> None:
         self.vllm_config = vllm_config
         self.model_config = vllm_config.model_config
@@ -46,6 +47,15 @@ class WorkerBase(ABC):
         self.prompt_adapter_config = vllm_config.prompt_adapter_config
         self.observability_config = vllm_config.observability_config
         self.kv_transfer_config = vllm_config.kv_transfer_config
+        self.local_rank = local_rank
+
+        if (self.cache_config.enable_kv_store):
+            self.cache_config.kv_store = KVBlockStore.from_configs(
+                self.cache_config, self.model_config, self.parallel_config,
+                torch.device(f"cuda:{self.local_rank}"))
+        self.kv_store = self.cache_config.kv_store
+        self.kv_store_manager = self.cache_config.kv_store_manager
+
 
     @abstractmethod
     def init_device(self) -> None:
@@ -242,38 +252,51 @@ class LocalOrDistributedWorkerBase(WorkerBase):
         raise NotImplementedError
 
     @abstractmethod
+    def execute_worker(self, worker_input: WorkerInput) -> None:
+        """
+        Process an execution request.
+        """
+        raise NotImplementedError
+
     def prepare_kv_store_meta(self,
                               is_prefill: Optional[bool],
                               incomplete_put_block_ids: torch.Tensor,
                               put_block_ids: torch.Tensor,
                               seq_g_list: List[SequenceGroupMetadata]) \
                                       -> KVStoreMeta:
-        """
-        Prepare the KVStoreMeta for the worker. This is called by the driver
-        worker to prepare the metadata for the KVStore.
-        """
-        pass
+        ret_incomplete_put_blocks = torch.Tensor()
+        ret_put_blocks_mapping = torch.Tensor()
+        ret_seq_g_ids = torch.Tensor()
+        if (self.local_rank == 0) and (self.kv_store_manager is not None):
+            self.kv_store_manager.is_prefill = is_prefill
+            (ret_incomplete_put_blocks, ret_put_blocks_mapping) = \
+                    self.kv_store_manager.get_put_blocks_mapping(
+                        incomplete_put_block_ids, put_block_ids)
+            self.kv_store_manager.is_prefill = False
+            if (is_prefill) and (ret_incomplete_put_blocks.numel() + \
+                    ret_put_blocks_mapping.numel() > 0):
+                # XXX: use first seq_id representing the seq_group id
+                seq_g_ids = [seq_g.get_first_seq_id() for seq_g in seq_g_list]
+                ret_seq_g_ids = torch.tensor(seq_g_ids,
+                                             device="cpu",
+                                             dtype=torch.int64).view(-1)
+        return KVStoreMeta(ret_incomplete_put_blocks, ret_put_blocks_mapping,
+                           ret_seq_g_ids)
 
-    @abstractmethod
-    def put_stream_sync(self) -> None:
-        """
-        Synchronize the stream.
-        """
-        raise NotImplementedError
+    def put_stream_sync(self):
+        if (self.kv_store is not None):
+            self.kv_store.put_stream_sync()
 
-    @abstractmethod
     def issue_blocks_copy(self, worker_input: WorkerInput) -> None:
-        """
-        Issue the copy of the blocks from CPU to GPU with the given indices.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def execute_worker(self, worker_input: WorkerInput) -> None:
-        """
-        Process an execution request.
-        """
-        raise NotImplementedError
+        if (self.kv_store is None):
+            return
+        kv_caches = (self.kv_cache[worker_input.virtual_engine]
+                     if self.kv_cache is not None else None)
+        self.kv_store.get_blocks(
+            BlockMappingFromCPU(worker_input.kv_store_block_mapping,
+                                worker_input.kv_store_block_offsets,
+                                worker_input.kv_store_block_req_ids),
+            kv_caches)
 
     def _get_worker_input_from_broadcast(
         self
